move landing section
