+++ src/components/Landing/Header/Header.css
@@ -1,5 +1,5 @@
 .header {
-  background-image: url('../assets/bg.webp');
+  background-image: url('../../../assets/bg.webp');
   height: 550px;
   background-size: 150vw;
   background-repeat: no-repeat;
