+++ src/components/Landing/Header/Header.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-unused-vars */
 import React from 'react';
-import '../styles/Header.css';
+import './Header.css';
 import { Button, Jumbotron } from 'react-bootstrap';
 import Container from 'react-bootstrap/Container';
 import {Link} from "react-router-dom";
