+++ src/components/Landing/LandingSection/LandingSection.js
@@ -1,10 +1,10 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import React from 'react';
-import '../styles/LandingSection.css';
+import './LandingSection.css';
 import { Col, Container, Row, Image } from 'react-bootstrap';
-import Img from '../assets/section-img.webp';
-import ClockImg from '../assets/clock.webp';
-import TalkingImg from '../assets/talking.webp';
+import Img from '../../../assets/section-img.webp';
+import ClockImg from '../../../assets/clock.webp';
+import TalkingImg from '../../../assets/talking.webp';
 
 function LandingSection() {
   return (
