+++ src/components/index.js
@@ -2,4 +2,6 @@ export * from './marginals';
 
 export { default as AddJournal} from './AddJournal/AddJournal';
 export { default as Login } from './Authentication/Login/Login';
-export { default as SignUp } from './Authentication/Signup/SignUp'
+export { default as SignUp } from './Authentication/Signup/SignUp';
+export { default as Header } from './Landing/Header/Header';
+export { default as LandingSection } from './Landing/LandingSection/LandingSection';
