+++ src/pages/Home.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from '../components/Header';
-import LandingSection from '../components/LandingSection';
-import "../styles/Header.css"
+import Header from '../components/Landing/Header/Header';
+import { LandingSection } from '../components';
+import "../components/Landing/Header/Header.css"
 
 function Home() {
   return (
